+++ tools/Customization/DevHome.Customization/Views/MainPage.xaml
@@ -1,12 +1,12 @@
 <!--  Copyright (c) Microsoft Corporation..  -->
 <!--  Licensed under the MIT License.  -->
 
-<pg:ToolPage
+<commonviews:ToolPage
     x:Class="DevHome.Customization.Views.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:behaviors="using:DevHome.Common.Behaviors"
-    xmlns:pg="using:DevHome.Common"
+    xmlns:commonviews="using:DevHome.Common.Views"
     xmlns:views="using:DevHome.Customization.Views"
     behaviors:NavigationViewHeaderBehavior.HeaderTemplate="{StaticResource BreadcrumbBarDataTemplate}"
     behaviors:NavigationViewHeaderBehavior.HeaderContext="{x:Bind ViewModel}">
@@ -22,4 +22,4 @@
             <views:MainPageView />
         </ScrollView>
     </Grid>
-</pg:ToolPage>
+</commonviews:ToolPage>
