+++ tools/Customization/DevHome.Customization/Views/MainPage.xaml.cs
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using DevHome.Common;
 using DevHome.Common.Extensions;
+using DevHome.Common.Views;
 using DevHome.Customization.ViewModels;
 using Microsoft.UI.Xaml;
