+++ tools/Dashboard/DevHome.Dashboard/Views/DashboardView.xaml
@@ -1,14 +1,13 @@
 <!-- Copyright (c) Microsoft Corporation.. -->
 <!-- Licensed under the MIT License. -->
 
-<pg:ToolPage
+<commonviews:ToolPage
     x:Class="DevHome.Dashboard.Views.DashboardView"
     x:Uid="DashboardPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:pg="using:DevHome.Common"
     xmlns:commonviews="using:DevHome.Common.Views"
     xmlns:vm="using:DevHome.Dashboard.ViewModels"
     xmlns:views="using:DevHome.Dashboard.Views"
@@ -151,4 +150,4 @@
             </Grid>
         </ScrollViewer>
     </Grid>
-</pg:ToolPage>
+</commonviews:ToolPage>
