+++ tools/Experiments/src/Views/TestExperimentPage.xaml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
-<pg:ToolPage
+<commonviews:ToolPage
     x:Class="DevHome.Experiments.Views.TestExperimentPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:DevHome.Experiments.Views"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:pg="using:DevHome.Common"
+    xmlns:commonviews="using:DevHome.Common.Views"
     mc:Ignorable="d">
 
     <Grid>
 
     </Grid>
-</pg:ToolPage>
+</commonviews:ToolPage>
