+++ tools/ExtensionLibrary/DevHome.ExtensionLibrary/Views/ExtensionLibraryView.xaml
@@ -1,11 +1,10 @@
 <!-- Copyright (c) Microsoft Corporation.. -->
 <!-- Licensed under the MIT License. -->
 
-<pg:ToolPage
+<commonviews:ToolPage
     x:Class="DevHome.ExtensionLibrary.Views.ExtensionLibraryView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:pg="using:DevHome.Common"
     xmlns:converters="using:CommunityToolkit.WinUI.Converters"
     xmlns:ctControls="using:CommunityToolkit.WinUI.Controls"
     xmlns:viewmodels="using:DevHome.ExtensionLibrary.ViewModels"
@@ -154,4 +153,4 @@
             </StackPanel>
         </ScrollViewer>
     </Grid>
-</pg:ToolPage>
+</commonviews:ToolPage>
