+++ tools/SampleTool/src/Views/SampleToolPage.xaml
@@ -1,11 +1,11 @@
-﻿<pg:ToolPage
+﻿<commonviews:ToolPage
     x:Class="Tools.SampleTool.Views.SampleToolPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:pg="using:DevHome.Common"
+    xmlns:commonviews="using:DevHome.Common.Views"
     mc:Ignorable="d">
 
     <TextBlock x:Uid="ContentArea" />
-</pg:ToolPage>
+</commonviews:ToolPage>
