+++ tools/SetupFlow/DevHome.SetupFlow/Views/SetupFlowPage.xaml.cs
@@ -1,8 +1,8 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using DevHome.Common;
 using DevHome.Common.Extensions;
+using DevHome.Common.Views;
 using DevHome.SetupFlow.ViewModels;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.UI.Xaml;
