+++ tools/Utilities/src/Views/UtilitiesMainPageView.xaml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<pg:ToolPage
+<commonviews:ToolPage
     x:Class="DevHome.Utilities.Views.UtilitiesMainPageView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -7,7 +7,7 @@
     xmlns:viewmodels="using:DevHome.Utilities.ViewModels"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:pg="using:DevHome.Common"
+    xmlns:commonviews="using:DevHome.Common.Views"
     xmlns:animations="using:CommunityToolkit.WinUI.Animations"
     mc:Ignorable="d">
 
@@ -41,4 +41,4 @@
             </ItemsRepeater>
         </Grid>
     </ScrollViewer>
-</pg:ToolPage>
+</commonviews:ToolPage>
