+++ tools/Utilities/src/Views/UtilitiesMainPageView.xaml.cs
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 
 using System.Windows.Input;
-using DevHome.Common;
 using DevHome.Common.Extensions;
+using DevHome.Common.Views;
 using DevHome.Utilities.ViewModels;
 using Microsoft.UI.Xaml;
